Fix SSR rendering by sharing the resolver context.
The resolver context sets up all the clients (fetch/jsonrpc) to interact
with DDF. When using the schema link, we need to ensure resolvers still
get clients or they won't be able to communicate with DDF. Those clients
also need access to the current HTTP request to pull auth info from
headers.
